add instanceof support
