+++ packages/config-transformer/packages/format-switcher/src/PhpParser/NodeFactory/CommonFactory.php
@@ -4,14 +4,27 @@ declare(strict_types=1);
 
 namespace Migrify\ConfigTransformer\FormatSwitcher\PhpParser\NodeFactory;
 
+use Migrify\ConfigTransformer\Naming\ClassNaming;
 use PhpParser\BuilderHelpers;
 use PhpParser\Node\Expr;
 use PhpParser\Node\Expr\BinaryOp\Concat;
+use PhpParser\Node\Expr\ClassConstFetch;
+use PhpParser\Node\Name;
 use PhpParser\Node\Scalar\MagicConst\Dir;
 use PhpParser\Node\Scalar\String_;
 
 final class CommonFactory
 {
+    /**
+     * @var ClassNaming
+     */
+    private $classNaming;
+
+    public function __construct(ClassNaming $classNaming)
+    {
+        $this->classNaming = $classNaming;
+    }
+
     public function createAbsoluteDirExpr($argument): Expr
     {
         if (is_string($argument)) {
@@ -26,5 +39,11 @@ final class CommonFactory
         }
 
         return $argumentValue;
+    }
+
+    public function createShortClassReference(string $className): ClassConstFetch
+    {
+        $shortClassName = $this->classNaming->getShortName($className);
+        return new ClassConstFetch(new Name($shortClassName), 'class');
     }
 }
